+++ operations/running.go
@@ -172,7 +172,7 @@ func LoadTaskWithEnvironment(packfile, theTask string, force bool) (bool, robot.
 		return true, config, todo, ""
 	}
 
-	label, _, err := htfs.NewEnvironment(config.CondaConfigFile(), config.Holozip(), true, force)
+	label, _, err := htfs.NewEnvironment(config.CondaConfigFile(), config.Holozip(), true, force, PullCatalog)
 	if err != nil {
 		pretty.Exit(4, "Error: %v", err)
 	}